requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiormq==6.8.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
billiard==4.2.1
blinker==1.8.2
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.5
filelock==3.16.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gotrue==2.8.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
hyperframe==6.0.1
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
kiwisolver==1.4.7
kombu==5.4.2
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
moviepy==1.0.3
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.1.0
mutagen==1.47.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai==1.47.1
openai-whisper @ git+https://github.com/openai/whisper.git@279133e3107392276dc509148da1f41bfb532c7e
opencv-python==4.10.0.84
packaging==24.1
pamqp==3.3.0
pandas==2.2.2
pika==1.3.2
pillow==10.4.0
postgrest==0.16.11
proglog==0.1.10
prompt_toolkit==3.0.47
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
realtime==2.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
sympy==1.13.3
ta==0.11.0
TA-Lib==0.4.32
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
yarl==1.12.1
yt-dlp==2024.8.6
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiormq==6.8.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
billiard==4.2.1
blinker==1.8.2
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.5
filelock==3.16.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gotrue==2.8.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
hyperframe==6.0.1
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
kiwisolver==1.4.7
kombu==5.4.2
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
moviepy==1.0.3
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.1.0
mutagen==1.47.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai==1.47.1
openai-whisper @ git+https://github.com/openai/whisper.git@279133e3107392276dc509148da1f41bfb532c7e
opencv-python==4.10.0.84
packaging==24.1
pamqp==3.3.0
pandas==2.2.2
pika==1.3.2
pillow==10.4.0
postgrest==0.16.11
proglog==0.1.10
prompt_toolkit==3.0.47
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
realtime==2.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
sympy==1.13.3
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
yarl==1.12.1
yt-dlp==2024.8.6
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiormq==6.8.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
billiard==4.2.1
blinker==1.8.2
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.5
filelock==3.16.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gotrue==2.8.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
hyperframe==6.0.1
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
kiwisolver==1.4.7
kombu==5.4.2
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
moviepy==1.0.3
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.1.0
mutagen==1.47.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai==1.47.1
openai-whisper @ git+https://github.com/openai/whisper.git@279133e3107392276dc509148da1f41bfb532c7e
opencv-python==4.10.0.84
packaging==24.1
pamqp==3.3.0
pandas==2.2.2
pika==1.3.2
pillow==10.4.0
postgrest==0.16.11
proglog==0.1.10
prompt_toolkit==3.0.47
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
realtime==2.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
sympy==1.13.3
ta==0.11.0
TA-Lib==0.4.32
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
yarl==1.12.1
yt-dlp==2024.8.6
